requirements.txt
seleniumbase>=4.46.0
requests>=2.32.5
psycopg2-binary==2.9.9
pydantic==2.5.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pandas==2.1.4
gunicorn==21.2.0
psutil>=6.0.0
memory_profiler>=0.61.0
